Step R24: DataAnnotation to Tier B, with its relational base (#56)
Moves DataAnnotation from ADR-009 Tier A (InMemory) to Tier B (SQLite) and
onto DataAnnotationRelationalTestBase, under the Group C policy.

The relational base adds only two tests, but the value of this move is the
store. The Tier A class carried six overrides that each replaced a round
trip with a metadata assertion, because InMemory enforces no store
constraint. On a real database three of the six run for real and pass:
ConcurrencyCheckAttribute_throws_if_value_in_database_changed and the two
RequiredAttribute ones.

The UseTransaction override is written in the same commit as the store
switch, per CLAUDE.md. The base routes five tests through
ExecuteWithStrategyInTransactionAsync -- four in the core base and one in
the relational one -- and the tell is the base's own transaction strategy,
not anything visible in the fixture.

The other three of the six are replaced by EF's own DataAnnotationSqliteTest
overrides: same bodies, same reasons on a different store. SQLite enforces
no column length (MaxLengthAttribute, StringLengthAttribute) and has no
rowversion (TimestampAttribute, EF issue #2195).

One test is left failing. The relational base's
Table_can_configure_TPT_with_Owned asserts over
context.Model.GetTableMappings(), which needs the relational model, and the
model under assertion is the client's -- which this provider does not build
relationally (M9). That is R23's JsonTypes finding at one-hundredth the
size: there the same assumption cost 92 tests and the base is not adoptable,
here it costs one, so the base is adopted and the test is left failing per
ADR-004. Exactly the distinction ADR-013's 2026-08-30 amendment draws.

Class goes 95 / 0 failed / 95 total to 97 total, 96 passed, 1 failed.
failed 70 -> 71, total 27024 -> 27026. Baseline files, names file and plan
updated in this commit.
