+++ test/known-failures.names.txt
@@ -9,6 +9,7 @@ InfoCarrier.Core.FunctionalTests.Sqlite.ComplexTypesTrackingInfoCarrierTest.Can_
 InfoCarrier.Core.FunctionalTests.Sqlite.ComplexTypesTrackingInfoCarrierTest.Can_track_entity_with_complex_property_bag_collections(state: Added, async: True)
 InfoCarrier.Core.FunctionalTests.Sqlite.CustomConvertersInfoCarrierTest.Collection_enum_as_string_Contains
 InfoCarrier.Core.FunctionalTests.Sqlite.CustomConvertersInfoCarrierTest.Composition_over_collection_of_complex_mapped_as_scalar
+InfoCarrier.Core.FunctionalTests.Sqlite.DataAnnotationInfoCarrierTest.Table_can_configure_TPT_with_Owned
 InfoCarrier.Core.FunctionalTests.Sqlite.EntitySplittingInfoCarrierTest.ExecuteDelete_throws_for_entity_splitting(async: False)
 InfoCarrier.Core.FunctionalTests.Sqlite.EntitySplittingInfoCarrierTest.ExecuteDelete_throws_for_entity_splitting(async: True)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.NorthwindSetOperationsQueryInfoCarrierTest.Collection_projection_before_set_operation_fails(async: False)
